+++ src/solver/task/solution.rs
@@ -29,6 +29,7 @@ pub enum SolveError {
     ExecutionDeadline,
     Canceled,
     TimeDeadline,
+    Internal,
 }
 
 #[derive(Debug, Display, Default, Clone, Copy, Encode, Decode, PartialEq, Eq)]